refactor: Move book search functionality and state to main App component
This change ensures that books can only belong to one shelf at a time and that books that are already shelved are not shown in the search results / available books section.
The moving of books to different shelves are now handled on success after success from the server. Previously it was optimistically moved which meant that we had to explictly
move the book back if the network request failed. To ensure that the user can't now move the book again before the network request is done, we hide the selector while it is moving.